Client/requirements.txt
pycrypto==2.6.1
behave==1.2.5
base58==0.2.5
msgpack-python==0.4.8